@@ -13,7 +13,7 @@ namespace FilterLists.Api.DependencyInjection.Extensions
             services.ConfigureCookiePolicy();
             services.AddMemoryCache();
             services.AddResponseCaching();
-            services.AddMvcCustom();
+            services.AddControllersCustom();
             services.AddRoutingCustom();
             services.AddApiVersioning();
         }
@@ -25,10 +25,10 @@ namespace FilterLists.Api.DependencyInjection.Extensions
                 opts.MinimumSameSitePolicy = SameSiteMode.None;
             });
 
-        private static void AddMvcCustom(this IServiceCollection services) =>
-            services.AddMvc()
-                    .SetCompatibilityVersion(CompatibilityVersion.Version_2_1)
-                    .AddJsonOptions(opts =>
+        private static void AddControllersCustom(this IServiceCollection services) =>
+            services.AddControllers()
+                    .SetCompatibilityVersion(CompatibilityVersion.Version_3_0)
+                    .AddNewtonsoftJson(opts =>
                     {
                         opts.SerializerSettings.NullValueHandling = NullValueHandling.Ignore;
                         opts.SerializerSettings.ContractResolver = new SkipEmptyContractResolver();
